parser-bot/bot/requirements.txt
python-telegram-bot[async]==20.3
httpx==0.24.1
python-dotenv==1.0.0